Reuse shared LAN discovery infrastructure
