+++ Basis/Packages/com.basis.server/BasisNetworkCore/BasisLanDiscoveryProtocol.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2c8fb92b8bca4ee3ad25dcfc2e2cc09a
