+++ Basis/Packages/com.basis.server/BasisNetworkCore/BasisLanMdnsTransport.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 684df842763b48b0890a4c1e8fd9c142
